v7
Added RP Images to be loaded. Right now, this looks at RP username and
uses that as the filename of the image. It assumes that the images are
in the extension object folder, which is probably not realistic. A
better approach might be to call these through the web - is this
possible?